Changed major code logic
Added comments, changed a lot of logic, removed counter and added a stopwatch instead. To calculate at the end of the race, how long time it took for each car, just for fun. Split up a lot of the code into separate classes to increase readability and some small fixes and changes here and there.