Created stopwatch and timespan to keep track of the race events instead of an int counter.
Changed stopwatch to public to be able to use it in other classes.

Fixed names of the renamed car properties.

Removed car.seconds references as we use eventTimer for that instead.

Added a writeline to print out "press enter" when all cars have finished. As it wasn't clear what we were wating for when race is over.